excel-autograder-backend/requirements.txt
Django~=4.2.5
psycopg2-binary~=2.9.7
colorama~=0.4.6
djangorestframework~=3.14.0
django-cors-headers~=4.2.0
Pillow~=10.0.1
uuid~=1.30